Remove the "nepyvo" series
It turns out that pyvo-data is not a good place for *all*
Python-related events -- only for the monthly meetups.

Nepyvo data was never shown on pyvo.cz; it was only in this
data repository.
Remove it to get rid of the special case when manipulating
the data.

The historical data will stay in the `archive/nepyvo`
branch of the repository.